First use image updates
Combined IBD info and connection settings, which also removed the connection category options. We decide to not include them because they are mostly relevant to mobile, which will not be supported for the time being.

Minor copy update to the storage screens for clarity (avoiding specific GB amounts as they can be confusing).