Research section
The research done at the beginning of this year would be good to include on the website itself. This PR aims to create a research section as well as include the report of the user interviews that were done. This would help so that it can be referred or linked to in the design docs and so that all research is public and in a single location.