Quick first draft of a features (comparison) table
Just want to get something going. It's a basic markdown table. Each row lists a feature. The name is a link to the respective place in the docs. The other two columns are to indicate support in the application and support in the QT application. Very simple.

If we like this format, we can properly gather the table content.

Closes #125